pyproject.toml
[project]
name = "automate-job-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.60.0",
    "google-generativeai>=0.8.6",
    "gspread>=6.2.1",
    "oauth2client>=4.1.3",
    "pandas>=2.3.3",
    "pypdf2>=3.0.1",
    "python-jobspy>=1.1.82",
]
